+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 ## 1.0.0 (2020-03-13)
 
-
 ### Features
 
-* initial release of `import-sort-style-module-alias` ([7f019b2](https://github.com/ifiokjr/import-sort-style-module-alias/commit/7f019b24818581f884d2d43b4b6b34c093bc7bde))
+- initial release of `import-sort-style-custom` ([7f019b2](https://github.com/ifiokjr/import-sort-style-custom/commit/7f019b24818581f884d2d43b4b6b34c093bc7bde))